Migrate item list ViewModel code to new one
