+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/library/CustomizeTabsDialog.kt
@@ -57,9 +57,9 @@ import sh.calvin.reorderable.rememberReorderableLazyListState
 
 @Composable
 fun CustomizeTabsDialog(
-    initialConfig: List<Pair<ItemListViewModel.Tab, Boolean>>,
+    initialConfig: List<Pair<LibraryTabsViewModel.Tab, Boolean>>,
     onDismissRequest: () -> Unit,
-    onConfirm: (List<Pair<ItemListViewModel.Tab, Boolean>>) -> Unit,
+    onConfirm: (List<Pair<LibraryTabsViewModel.Tab, Boolean>>) -> Unit,
 ) {
     Dialog(onDismissRequest = onDismissRequest) {
         Card(
@@ -87,8 +87,8 @@ fun CustomizeTabsDialog(
 
 @Composable
 private fun TabsCustomizeList(
-    initialConfig: List<Pair<ItemListViewModel.Tab, Boolean>>,
-    onDone: (List<Pair<ItemListViewModel.Tab, Boolean>>) -> Unit,
+    initialConfig: List<Pair<LibraryTabsViewModel.Tab, Boolean>>,
+    onDone: (List<Pair<LibraryTabsViewModel.Tab, Boolean>>) -> Unit,
 ) {
     val plateShape = RoundedCornerShape(12.dp)
     var items by remember { mutableStateOf(initialConfig) }
@@ -171,22 +171,22 @@ private fun TabsCustomizeList(
     }
 }
 
-private fun ItemListViewModel.Tab.labelRes(): StringResource = when (this) {
-    ItemListViewModel.Tab.ARTISTS -> Res.string.media_type_artists
-    ItemListViewModel.Tab.ALBUMS -> Res.string.media_type_albums
-    ItemListViewModel.Tab.TRACKS -> Res.string.media_type_tracks
-    ItemListViewModel.Tab.PLAYLISTS -> Res.string.media_type_playlists
-    ItemListViewModel.Tab.AUDIOBOOKS -> Res.string.media_type_audiobooks
-    ItemListViewModel.Tab.PODCASTS -> Res.string.media_type_podcasts
-    ItemListViewModel.Tab.RADIOS -> Res.string.media_type_radio
-    ItemListViewModel.Tab.GENRES -> Res.string.media_type_genres
+private fun LibraryTabsViewModel.Tab.labelRes(): StringResource = when (this) {
+    LibraryTabsViewModel.Tab.ARTISTS -> Res.string.media_type_artists
+    LibraryTabsViewModel.Tab.ALBUMS -> Res.string.media_type_albums
+    LibraryTabsViewModel.Tab.TRACKS -> Res.string.media_type_tracks
+    LibraryTabsViewModel.Tab.PLAYLISTS -> Res.string.media_type_playlists
+    LibraryTabsViewModel.Tab.AUDIOBOOKS -> Res.string.media_type_audiobooks
+    LibraryTabsViewModel.Tab.PODCASTS -> Res.string.media_type_podcasts
+    LibraryTabsViewModel.Tab.RADIOS -> Res.string.media_type_radio
+    LibraryTabsViewModel.Tab.GENRES -> Res.string.media_type_genres
 }
 
 @Preview
 @Composable
 private fun PreviewCustomizeTabsDialog() {
     CustomizeTabsDialog(
-        initialConfig = ItemListViewModel.Tab.entries.mapIndexed { i, t -> t to (i < 5) },
+        initialConfig = LibraryTabsViewModel.Tab.entries.mapIndexed { i, t -> t to (i < 5) },
         onDismissRequest = {},
         onConfirm = {},
     )
